backend/requirements.txt
aniso8601==9.0.1
aws-wsgi==0.2.7
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Cors==4.0.1
Flask-RESTful==0.3.10
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
tinydb==4.8.0
urllib3==2.2.2
Werkzeug==3.0.3
